refactor(testing): detach test infrastructure from runtime
- add the repo rule for detached test infrastructure
- replace runtime test seams with factories, unregister handles, and module-level test mocks
- split the content-script runtime out of the WXT entrypoint
