+++ apps/extension/src/lib/events.test.ts
@@ -1,8 +1,13 @@
 import { describe, expect, it } from 'vitest';
 import type { CorrectionEvent } from '@movar/events';
-import { __internal, logCorrection, logCorrections } from './events';
-
-const { prune, getEvents, EVENT_TTL_MS, MAX_EVENTS } = __internal;
+import {
+  EVENT_TTL_MS,
+  MAX_EVENTS,
+  getCorrectionEvents,
+  logCorrection,
+  logCorrections,
+  pruneCorrectionEvents,
+} from './events';
 
 function event(timestamp: number): CorrectionEvent {
   return {
@@ -19,7 +24,7 @@ describe('events prune', () => {
 
   it('keeps events younger than the TTL', () => {
     const events = [event(NOW - 100), event(NOW - 1000)];
-    expect(prune(events, NOW)).toHaveLength(2);
+    expect(pruneCorrectionEvents(events, NOW)).toHaveLength(2);
   });
 
   it('drops events older than the TTL', () => {
@@ -27,7 +32,7 @@ describe('events prune', () => {
       event(NOW - EVENT_TTL_MS - 1), // just over the floor
       event(NOW - 1000),
     ];
-    const result = prune(events, NOW);
+    const result = pruneCorrectionEvents(events, NOW);
     expect(result).toHaveLength(1);
     expect(result[0]!.timestamp).toBe(NOW - 1000);
   });
@@ -35,12 +40,12 @@ describe('events prune', () => {
   it('keeps an event landing exactly at the TTL cutoff', () => {
     // Inclusive at the floor — `>= cutoff` not `> cutoff`.
     const events = [event(NOW - EVENT_TTL_MS)];
-    expect(prune(events, NOW)).toHaveLength(1);
+    expect(pruneCorrectionEvents(events, NOW)).toHaveLength(1);
   });
 
   it('respects MAX_EVENTS even when all are within the TTL', () => {
     const events = Array.from({ length: MAX_EVENTS + 50 }, (_, i) => event(NOW - i));
-    const result = prune(events, NOW);
+    const result = pruneCorrectionEvents(events, NOW);
     expect(result).toHaveLength(MAX_EVENTS);
   });
 
@@ -50,21 +55,21 @@ describe('events prune', () => {
     const events: CorrectionEvent[] = [];
     for (let i = 0; i < 5; i++) events.push(event(NOW - EVENT_TTL_MS - 1000));
     for (let i = MAX_EVENTS + 3; i > 0; i--) events.push(event(NOW - i));
-    const result = prune(events, NOW);
+    const result = pruneCorrectionEvents(events, NOW);
     expect(result).toHaveLength(MAX_EVENTS);
     expect(result.at(-1)!.timestamp).toBe(NOW - 1);
   });
 
   it('returns a new array (does not mutate input)', () => {
     const events = [event(NOW)];
-    const result = prune(events, NOW);
+    const result = pruneCorrectionEvents(events, NOW);
     expect(result).not.toBe(events);
   });
 });
 
 // Storage isn't reset between tests, so assert on the DELTA rather than absolute
 // counts (the cap test, which IS absolute, runs last).
-const count = async (): Promise<number> => (await getEvents()).length;
+const count = async (): Promise<number> => (await getCorrectionEvents()).length;
 
 describe('logCorrections', () => {
   it('appends a whole batch in one write', async () => {
