+++ apps/extension/src/lib/events.ts
@@ -3,16 +3,19 @@ import type { CorrectionEvent } from '@movar/events';
 import { DAY_MS } from './time';
 
 const EVENTS_KEY = 'movar:events';
-const MAX_EVENTS = 1000;
+export const MAX_EVENTS = 1000;
 /** Time-based retention floor. The 1,000-entry cap is the privacy-policy
  *  commitment; this TTL is a second pruning axis for users whose corrections
  *  ramp up slowly (a few per week). Without it a months-old event would sit
  *  in storage waiting for the cap to kick in. 30 days comfortably covers
  *  "today" + "this week" + "this month" UI surfaces. */
 const RETENTION_DAYS = 30;
-const EVENT_TTL_MS = RETENTION_DAYS * DAY_MS;
+export const EVENT_TTL_MS = RETENTION_DAYS * DAY_MS;
 
-function prune(events: readonly CorrectionEvent[], now: number): CorrectionEvent[] {
+export function pruneCorrectionEvents(
+  events: readonly CorrectionEvent[],
+  now: number,
+): CorrectionEvent[] {
   const cutoff = now - EVENT_TTL_MS;
   // Apply the time cutoff first, then the count cap — flipping the order
   // could let the cap drop fresh events behind a wall of stale ones if the
@@ -21,7 +24,7 @@ function prune(events: readonly CorrectionEvent[], now: number): CorrectionEvent
   return fresh.length > MAX_EVENTS ? fresh.slice(-MAX_EVENTS) : [...fresh];
 }
 
-async function getEvents(): Promise<CorrectionEvent[]> {
+export async function getCorrectionEvents(): Promise<CorrectionEvent[]> {
   const stored = await browser.storage.local.get(EVENTS_KEY);
   return (stored[EVENTS_KEY] as CorrectionEvent[] | undefined) ?? [];
 }
@@ -38,15 +41,11 @@ async function getEvents(): Promise<CorrectionEvent[]> {
  */
 export async function logCorrections(events: readonly CorrectionEvent[]): Promise<void> {
   if (events.length === 0) return;
-  const stored = await getEvents();
+  const stored = await getCorrectionEvents();
   stored.push(...events);
-  await browser.storage.local.set({ [EVENTS_KEY]: prune(stored, Date.now()) });
+  await browser.storage.local.set({ [EVENTS_KEY]: pruneCorrectionEvents(stored, Date.now()) });
 }
 
 export async function logCorrection(event: CorrectionEvent): Promise<void> {
   await logCorrections([event]);
 }
-
-/** Test seam — exposes the pruning policy for unit tests without forcing
- *  them through `chrome.storage.local`. Not part of the public runtime API. */
-export const __internal = { prune, getEvents, EVENT_TTL_MS, MAX_EVENTS };
