+++ packages/lang-detect/src/engines/chrome-ai.ts
@@ -45,69 +45,66 @@ interface LanguageDetectorApi {
   create(): Promise<LanguageDetectorSession>;
 }
 
-/** Module-scoped cache: once we've confirmed the API is available we keep
- *  saying so for the content-script lifetime; once we've confirmed it's not,
- *  we keep skipping. The orchestrator runs many times per tab, so per-call
- *  availability() calls would burn a meaningful chunk of the 150 ms budget. */
-let cachedAvailability: boolean | null = null;
-let cachedSession: LanguageDetectorSession | null = null;
-
 function getApi(): LanguageDetectorApi | null {
   const globalRef = globalThis as unknown as { LanguageDetector?: LanguageDetectorApi };
   return globalRef.LanguageDetector ?? null;
 }
 
-async function checkAvailability(): Promise<boolean> {
-  if (cachedAvailability !== null) return cachedAvailability;
-  const api = getApi();
-  if (!api) {
-    cachedAvailability = false;
-    return false;
-  }
-  const state = await api.availability();
-  cachedAvailability = state === 'available';
-  return cachedAvailability;
-}
+export function createChromeAiEngine(): LanguageDetectionEngine {
+  /** Instance cache: once we've confirmed the API is available we keep saying
+   *  so for the content-script lifetime; once we've confirmed it's not, we
+   *  keep skipping. The orchestrator runs many times per tab, so per-call
+   *  availability() calls would burn a meaningful chunk of the 150 ms budget. */
+  let cachedAvailability: boolean | null = null;
+  let cachedSession: LanguageDetectorSession | null = null;
 
-async function getSession(): Promise<LanguageDetectorSession> {
-  if (cachedSession) return cachedSession;
-  const api = getApi();
-  if (!api) throw new Error('chrome-ai: LanguageDetector API missing');
-  cachedSession = await api.create();
-  return cachedSession;
-}
-
-/** Test-only: reset the module-scoped state so each test starts cold. Exported
- *  with a double-underscore prefix to signal "not part of the public API". */
-export function __resetChromeAiCacheForTests(): void {
-  cachedAvailability = null;
-  cachedSession = null;
-}
-
-export const chromeAiEngine: LanguageDetectionEngine = {
-  id: ENGINE_ID,
-  // eslint-disable-next-line sonarjs/function-return-type -- intentional boolean | Promise<boolean> per the LanguageDetectionEngine.isAvailable contract: the cached/no-API fast paths return synchronously so the orchestrator skips without a promise round-trip; only the live availability() check is async
-  isAvailable(): boolean | Promise<boolean> {
+  async function checkAvailability(): Promise<boolean> {
     if (cachedAvailability !== null) return cachedAvailability;
-    // No API at all — return false synchronously so the orchestrator can skip
-    // without paying for a promise round-trip in the common (non-Chrome) case.
-    if (!getApi()) {
+    const api = getApi();
+    if (!api) {
       cachedAvailability = false;
       return false;
     }
-    return checkAvailability();
-  },
-  async detect(text, ctx: DetectContext): Promise<DetectedLanguage | null> {
-    const session = await getSession();
-    const sample = text.slice(0, ctx.maxChars ?? DEFAULT_MAX_CHARS);
-    const results = await session.detect(sample);
-    const top = results[0];
-    if (!top) return null;
-    if (top.confidence < CONFIDENCE_THRESHOLD) return null;
-    return {
-      language: top.detectedLanguage,
-      confidence: top.confidence,
-      engine: ENGINE_ID,
-    };
-  },
-};
+    const state = await api.availability();
+    cachedAvailability = state === 'available';
+    return cachedAvailability;
+  }
+
+  async function getSession(): Promise<LanguageDetectorSession> {
+    if (cachedSession) return cachedSession;
+    const api = getApi();
+    if (!api) throw new Error('chrome-ai: LanguageDetector API missing');
+    cachedSession = await api.create();
+    return cachedSession;
+  }
+
+  return {
+    id: ENGINE_ID,
+    // eslint-disable-next-line sonarjs/function-return-type -- intentional boolean | Promise<boolean> per the LanguageDetectionEngine.isAvailable contract: the cached/no-API fast paths return synchronously so the orchestrator skips without a promise round-trip; only the live availability() check is async
+    isAvailable(): boolean | Promise<boolean> {
+      if (cachedAvailability !== null) return cachedAvailability;
+      // No API at all — return false synchronously so the orchestrator can skip
+      // without paying for a promise round-trip in the common (non-Chrome) case.
+      if (!getApi()) {
+        cachedAvailability = false;
+        return false;
+      }
+      return checkAvailability();
+    },
+    async detect(text, ctx: DetectContext): Promise<DetectedLanguage | null> {
+      const session = await getSession();
+      const sample = text.slice(0, ctx.maxChars ?? DEFAULT_MAX_CHARS);
+      const results = await session.detect(sample);
+      const top = results[0];
+      if (!top) return null;
+      if (top.confidence < CONFIDENCE_THRESHOLD) return null;
+      return {
+        language: top.detectedLanguage,
+        confidence: top.confidence,
+        engine: ENGINE_ID,
+      };
+    },
+  };
+}
+
+export const chromeAiEngine: LanguageDetectionEngine = createChromeAiEngine();
